+++ src/reverseIterate.js
@@ -1,25 +1,26 @@
 // @flow strict
 
 /*::
-import type {ImmutableTree} from './types.js';
+import type {ImmutableTree, Stack} from './types.js';
 import invariant from './invariant.js';
 */
 
 export default function* reverseIterate/*:: <T> */(
   tree/*: ImmutableTree<T> */,
 )/*: Generator<T, void, void> */ {
-  const stack = [];
+  let stack/*: Stack<T> | null */ = null;
   let cursor/*: ImmutableTree<T> */ = tree;
   do {
     while (cursor.size !== 0) {
-      stack.push(cursor);
+      stack = {node: cursor, next: stack};
       cursor = cursor.right;
     }
-    const next = stack.pop();
-    if (next) {
-      yield next.value;
-      /*:: invariant(next != null); */
-      cursor = next.left;
+    if (stack !== null) {
+      const node = stack.node;
+      yield node.value;
+      /*:: invariant(node != null && stack != null); */
+      cursor = node.left;
+      stack = stack.next;
     }
-  } while (stack.length || cursor.size !== 0);
+  } while (stack !== null || cursor.size !== 0);
 }